+++ packages/webpack-common/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @webpack-blocks/webpack-common - Changelog
 
+## Next release
+
+- Added `parseVersion()`
+
 ## 0.3.0
 
 Initial release.